requirements.txt
flet
pycryptodome